move findExistingIdentityId to its own file
