+++ handlers/identity-backfill/src/main/scala/com/gu/identityBackfill/PreReqCheck.scala
@@ -1,8 +1,5 @@
 package com.gu.identityBackfill
 
-import com.gu.identity.GetByEmail.IdentityAccount
-import com.gu.identity.{GetByEmail, GetByIdentityId}
-import com.gu.identity.GetByIdentityId.IdentityUser
 import com.gu.identityBackfill.TypeConvert._
 import com.gu.identityBackfill.Types._
 import com.gu.identityBackfill.salesforce.UpdateSalesforceIdentityId.IdentityId
@@ -13,7 +10,7 @@ import com.gu.util.apigateway.ApiGatewayResponse
 import com.gu.util.reader.Types.ApiGatewayOp._
 import com.gu.util.reader.Types._
 import com.gu.util.resthttp.LazyClientFailableOp
-import com.gu.util.resthttp.Types.{ClientFailableOp, ClientFailure, ClientSuccess, NotFound}
+import com.gu.util.resthttp.Types.ClientFailableOp
 
 object PreReqCheck {
 
@@ -32,28 +29,6 @@ object PreReqCheck {
       _ <- maybeExistingIdentityId.map(noZuoraAccountsForIdentityId).getOrElse(ContinueProcessing(()))
       _ <- syncableSFToIdentity(zuoraAccountForEmail.sfContactId).value.toApiGatewayOp("load SF contact").flatMap(identity)
     } yield PreReqResult(zuoraAccountForEmail.accountId, zuoraAccountForEmail.sfContactId, maybeExistingIdentityId)
-  }
-
-  def findExistingIdentityId(
-    getByEmail: EmailAddress => ClientFailableOp[GetByEmail.IdentityAccount],
-    getByIdentityId: IdentityId => ClientFailableOp[GetByIdentityId.IdentityUser]
-  )(emailAddress: EmailAddress): ApiGatewayOp[Option[IdentityId]] = {
-
-    def continueIfNoPassword(identityId: IdentityId) = {
-      getByIdentityId(identityId) match {
-        case ClientSuccess(IdentityUser(_, false)) => ContinueProcessing(Some(identityId))
-        case _ => ReturnWithResponse(ApiGatewayResponse.notFound(s"identity email not validated but password is set $identityId"))
-      }
-    }
-
-    val result = getByEmail(emailAddress) match {
-      case ClientSuccess(IdentityAccount(identityId, true)) => ContinueProcessing(Some(identityId))
-      case ClientSuccess(IdentityAccount(identityId, false)) => continueIfNoPassword(identityId)
-      case NotFound(_) => ContinueProcessing(None)
-      case other: ClientFailure => ReturnWithResponse(ApiGatewayResponse.internalServerError(other.toString))
-    }
-
-    result.withLogging("GetByEmail")
   }
 
   def noZuoraAccountsForIdentityId(
