+++ handlers/identity-backfill/src/test/scala/com/gu/identityBackfill/PreReqCheckTest.scala
@@ -2,7 +2,6 @@ package com.gu.identityBackfill
 
 import com.gu.identity.GetByEmail
 import com.gu.identity.GetByEmail.IdentityAccount
-import com.gu.identity.GetByIdentityId.IdentityUser
 import com.gu.identityBackfill.PreReqCheck.PreReqResult
 import com.gu.identityBackfill.Types._
 import com.gu.identityBackfill.salesforce.UpdateSalesforceIdentityId.IdentityId
@@ -11,7 +10,7 @@ import com.gu.salesforce.TypesForSFEffectsData.SFContactId
 import com.gu.util.apigateway.ApiGatewayResponse
 import com.gu.util.reader.Types.ApiGatewayOp.{ContinueProcessing, ReturnWithResponse}
 import com.gu.util.resthttp.LazyClientFailableOp
-import com.gu.util.resthttp.Types.{ClientFailableOp, ClientFailure, ClientSuccess, GenericError, NotFound}
+import com.gu.util.resthttp.Types.{ClientFailableOp, ClientSuccess}
 import org.scalatest.{FlatSpec, Matchers}
 
 class PreReqCheckTest extends FlatSpec with Matchers {
@@ -129,47 +128,5 @@ class PreReqCheckTest extends FlatSpec with Matchers {
   "acceptable reader type" should "not allow a combination of valid and invalid" in {
     val readerTypes = List(ReaderType.ReaderTypeValue("Direct"), ReaderType.ReaderTypeValue("Gift"))
     PreReqCheck.acceptableReaderType(ClientSuccess(readerTypes)).toDisjunction.leftMap(_.statusCode) should be(scalaz.-\/("404"))
-  }
-
-  "findExistingIdentityId" should "continue processing with identity id for existing validated account" in {
-    PreReqCheck.findExistingIdentityId(
-      _ => ClientSuccess(IdentityAccount(IdentityId("100"), isUserEmailValidated = true)),
-      _ => fail("Should not be called")
-    )(EmailAddress("email@email.email")) should be(ContinueProcessing(Some(IdentityId("100"))))
-  }
-
-  "findExistingIdentityId" should "continue processing with identity id for existing unvalidated account with no password" in {
-    PreReqCheck.findExistingIdentityId(
-      _ => ClientSuccess(IdentityAccount(IdentityId("100"), isUserEmailValidated = false)),
-      _ => ClientSuccess(IdentityUser(IdentityId("100"), hasPassword = false))
-    )(EmailAddress("email@email.email")) should be(ContinueProcessing(Some(IdentityId("100"))))
-  }
-
-  "findExistingIdentityId" should "continue processing for not found identity user" in {
-    PreReqCheck.findExistingIdentityId(
-      _ => NotFound("not found"),
-      _ => fail("should not be called")
-    )(EmailAddress("email@email.email")) should be(ContinueProcessing(None))
-  }
-
-  "findExistingIdentityId" should "ReturnWithResponse for unvalidated account with password" in {
-    PreReqCheck.findExistingIdentityId(
-      _ => ClientSuccess(IdentityAccount(IdentityId("100"), isUserEmailValidated = false)),
-      _ => ClientSuccess(IdentityUser(IdentityId("100"), hasPassword = true))
-    )(EmailAddress("email@email.email")) should be(ReturnWithResponse(ApiGatewayResponse.notFound(s"identity email not validated but password is set IdentityId(100)")))
-  }
-
-  "findExistingIdentityId" should "ReturnWithResponse for unexpected identity response" in {
-    PreReqCheck.findExistingIdentityId(
-      _ => GenericError("error"),
-      _ => fail("should not be called")
-    )(EmailAddress("email@email.email")) should be(ReturnWithResponse(ApiGatewayResponse.internalServerError("error")))
-  }
-
-  "findExistingIdentityId" should "ReturnWithResponse for unexpected identity response for get by id" in {
-    PreReqCheck.findExistingIdentityId(
-      _ => ClientSuccess(IdentityAccount(IdentityId("100"), isUserEmailValidated = false)),
-      _ => GenericError("error"),
-    )(EmailAddress("email@email.email")) should be(ReturnWithResponse(ApiGatewayResponse.notFound(s"identity email not validated but password is set IdentityId(100)")))
   }
 }
